Provisioning: Refactor test endpoint
